requirements.txt
fastapi 
uvicorn 
python-dotenv 
openai
supabase    
python-socketio
aiohttp